Prevent double add or delete from bookshelf

+++ openreader/books.py
@@ -50,24 +50,43 @@ def bookshelf():
 	if not (action in ["add", "delete"] and book_id and user_id):
 		abort(400) # client error: bad request
 
-	if action == "delete":
+	prev_bookmark =  db.execute(
+		"SELECT book_id FROM bookshelf WHERE book_id = ? AND user_id = ?",
+		(book_id, user_id)
+	).fetchone()
+
+	if action == "delete" and prev_bookmark is not None:
 		db.execute(
 			"DELETE FROM bookshelf WHERE book_id = ? AND user_id = ?",
 			(book_id, user_id)
 		)
-	if action == "add":
+
+	if action == "add" and prev_bookmark is None:
 		db.execute(
 			"INSERT INTO bookshelf (book_id, user_id) VALUES (?, ?)",
 			(book_id, user_id)
 		)
+
 	db.commit()
 	return redirect(url_for("books.bookshelf"))
 
 
 @bp.route("/book/<int:id>")
 def info(id):
+	db = get_db()
 	res = readCatalog("/{}".format(id))
-	return render_template("bookInfo.html", book=res)
+	logged_in = g.user and g.user["id"]
+	bookmarked = None
+
+	if (logged_in): # if logged in, check if the user has a bookmark already
+		bookmarked =  db.execute(
+			"SELECT book_id FROM bookshelf WHERE book_id = ? AND user_id = ?",
+			(id, g.user["id"])
+		).fetchone() is not None
+	else:
+		bookmarked = False
+
+	return render_template("bookInfo.html", book=res, bookmarked=bookmarked)
 
 
 @bp.route("/book/<int:id>/read")
